Puts things how they should be
This is a much more  reasonable push, especially since I only really tested it with a couple sites, and none of the api ones, I will make sure to bring the gui up to functionality anywhere it is broken when I have time,  it was a very rushed job.
