+++ Build Release.bat
@@ -1,0 +1,1 @@
+conda activate pyinstaller && pyinstaller --paths Z:\Projects\Python\NN-Downloader\.env\Lib\site-packages --hidden-import requests --hidden-import inquirer --hidden-import alive_progress --hidden-import termcolor --hidden-import xmltodict --add-data="Z:\Projects\Python\NN-Downloader\.env\Lib\site-packages\grapheme\data\*;grapheme/data/" --onefile --icon "icon.ico" --console --name "NN-Downloader" --upx-dir "Z:\Projects\Python\### UPX ###" main.py
