+++ Experimental_Gui/.github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,38 @@
+---
+name: Bug report
+about: Create a bug report to help me solve the issue.
+title: "[BUG] "
+labels: bug, enhancement
+assignees: Official-Husko
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Desktop (please complete the following information):**
+ - OS: [e.g. Windows]
+ - Version [e.g. 10]
+
+**Used Parameters and Settings (please complete the following information):**
+- Site: [e.g. e621]
+- Tool Version: [e.g. ]
+- Tags/Link: [e.g. wolf female sfw]
+
+**Additional context**
+Add any other context about the problem here.
+
+### ***Please attach the runtime.log***
