+++ Experimental_Gui/__init__.py
@@ -1,0 +1,9 @@
+"""NN-Downloader - Modern async downloader with GUI support"""
+
+__version__ = "2.0.0"
+__author__ = "Official-Husko"
+
+from .core import *
+from .gui import *
+from .sites import *
+from .utils import *
