--- config.json
@@ -1,73 +1,0 @@
-{
-  "version": "2.0.0",
-  "check_for_updates": true,
-  "theme": "dark",
-  "log_level": "INFO",
-  "sites": {
-    "e621": {
-      "enabled": true,
-      "rate_limit": 5,
-      "credentials": null
-    },
-    "e6ai": {
-      "enabled": true,
-      "rate_limit": 5,
-      "credentials": null
-    },
-    "e926": {
-      "enabled": true,
-      "rate_limit": 5,
-      "credentials": null
-    },
-    "furbooru": {
-      "enabled": true,
-      "rate_limit": 5,
-      "credentials": null
-    },
-    "rule34": {
-      "enabled": true,
-      "rate_limit": 5,
-      "credentials": null
-    },
-    "luscious": {
-      "enabled": true,
-      "rate_limit": 5,
-      "credentials": null
-    },
-    "multporn": {
-      "enabled": true,
-      "rate_limit": 5,
-      "credentials": null
-    },
-    "yiffer": {
-      "enabled": true,
-      "rate_limit": 5,
-      "credentials": null
-    }
-  },
-  "proxy": {
-    "enabled": false,
-    "test_on_startup": true,
-    "timeout": 10,
-    "max_retries": 3
-  },
-  "download": {
-    "concurrent_downloads": 3,
-    "chunk_size": 8192,
-    "timeout": 30,
-    "retry_attempts": 3,
-    "preserve_metadata": true,
-    "ai_training_mode": false,
-    "one_time_download": true,
-    "output_directory": "media"
-  },
-  "blacklisted_tags": [
-    "gore",
-    "scat",
-    "watersports"
-  ],
-  "blacklisted_formats": [
-    "swf",
-    "webm"
-  ]
-}
