+++ modules/__init__.py
@@ -1,0 +1,14 @@
+from .proxyScraper import ProxyScraper
+from .configManager import Config_Manager
+from .auto_update import AutoUpdate
+from .logger import Logger
+from .pretty_print import *
+
+
+# Here are all modules for the sites that are supported
+from .e6systems import E6System
+from .rule34 import RULE34
+from .furbooru import FURBOORU
+from .multporn import Multporn
+from .yiffer import Yiffer
+from .luscious import Luscious
