+++ modules/pretty_print.py
@@ -1,0 +1,7 @@
+from termcolor import colored
+
+major_error = str(colored(f"[{colored('!!!', 'red')}]"))
+warning = str(colored(f"[{colored('!', 'yellow')}]"))
+error = str(colored(f"[{colored('!', 'red')}]"))
+info = str(colored(f"[{colored('i', 'light_blue')}]"))
+ok = str(colored(f"[{colored('+', 'green')}]"))
