+++ requirements.txt
@@ -1,13 +1,18 @@
-# NN-Downloader v2.0 - Essential Dependencies Only
-
-# Core async dependencies (required for downloads)
-aiohttp>=3.9.0
-aiofiles>=24.1.0
-
-# GUI dependencies (tkinter is built into Python)
-# No additional GUI dependencies needed
-
-# Optional dependencies (uncomment if needed)
-# requests>=2.32.0          # For compatibility with old modules
-# termcolor>=2.4.0          # For colored console output  
-# alive-progress>=3.1.5     # For progress bars in console mode
+about-time==4.2.1
+alive-progress==3.1.4
+ansicon==1.89.0
+blessed==1.20.0
+certifi==2023.7.22
+charset-normalizer==3.3.0
+grapheme==0.6.0
+idna==3.7
+inquirer==3.1.3
+jinxed==1.2.0
+python-editor==1.0.4
+readchar==4.0.5
+requests==2.32.0
+six==1.16.0
+termcolor==2.3.0
+urllib3==2.2.2
+wcwidth==0.2.8
+xmltodict==0.13.0
